// src/app/shared/components/app-dynamic-form-builder/models/field-builder.type.ts
export type FieldBuilderType =
    | 'password'
    | 'email'
    | 'number'
    | 'tel'
    | 'url'
    | 'time'
    | 'range'
    | 'hidden'
    | 'submit'
    | 'reset'
    | "text"
    | "dropdown"
    | "checkbox"
    | "radio"
    | "file"
    | "date"
    | "multiselect"
    | "multiselect-checkbox"
    | "autocomplete";